+++ package.json
@@ -9,7 +9,7 @@
     "lint:fix": "npx eslint --fix .",
     "format": "npx prettier --write \"**/*.+(js|jsx|json|yml|yaml|css|md)\"",
     "knex": "npx knex --knexfile config/knexfile.js",
-    "tests": "env NODE_ENV=test npx jest --forceExit",
+    "tests": "npx jest --forceExit",
     "coverage": "CI=true npx jest --coverage --detectOpenHandles --forceExit"
   },
   "jest": {
@@ -21,8 +21,7 @@
       "!**/__tests__/**.test.js",
       "!**/coverage/**",
       "!**/apidoc/**"
-    ],
-    "modulePathIgnorePatterns": ["./__tests__/integrationtests"]
+    ]
   },
   "apidoc": {
     "title": "Labs API Starter Scaffolding"
